requirements.txt
requests
jsonpath-ng
lxml
jsonschema
google-auth
faker
pytest-order
pytest-dependency
